+++ camunda-intellij-plugin/build.gradle
@@ -42,10 +42,6 @@ intellij {
     plugins = intellijPlatformPlugins
 }
 
-runIde {
-    jvmArgs '-Xmx2G'
-}
-
 runPluginVerifier {
     ideVersions = ["IU-2019.3.5"]
 }
